backend/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-seed==0.2.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Faker==4.16.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
